[PLAT-6779] Added method to BusinessDayDateUtils
